requirements.txt
audioread==3.0.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numba==0.60.0
numpy==1.23.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
prettytable==3.10.2
psutil==6.0.0
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.13.1
Send2Trash==1.8.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.13.0
threadpoolctl==3.5.0
torch==2.3.1
torchaudio==2.3.1
torchlibrosa==0.1.0
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
yamlloader==1.4.1
zipp==3.19.2
